Start a routing script that ties together others
I want to be able to invoke something like `ja journal ...` when I
ssh into the machine to post from my phone, or `ja status` to see
a health check and some stats. All of that is doable if I write
some good bash scripts and throw them in /usr/local/bin and tie
them together with a short router script. This is the start of that
effort.