hugo buildFuture true
Because my server and a lot of posting environments are in the British timezone,
notes and blog posts will end up with a date that looks to be an hour in the
future to the builder (UTC vs BST). To solve that I just publish all posts
infinitely into the future. If at some point I want to be able to schedule posts
for publishing I'll revisit this.